refactored valid type checking; used it to resolve #15
